requirements.txt
asyncio
websockets~=13.1
deepface~=0.0.93
tensorflow-cpu
numpy~=2.0.2
Pillow~=11.0.0
scikit-learn~=1.5.2
python-dotenv~=1.0.1
supabase~=2.10.0
tf-keras
python-docx
openpyxl~=3.1.5